refactor(solana): delete dead fhe_eval derivation surface (fhevm-internal#1853 W1+W2)
- drop computed_eval_handle_for_current_slot{,_with_chain_id},
  computed_unary_handle, computed_eval_rand_bounded_seed (zero callers
  repo-wide, incl. host-listener, proof service, SDK)
- deduplicate the two inlined copies of finish_computed_handle in the
  rand handle derivations
- fold handles.rs (pure forwarders) into methods on EvalHandleContext
  in walk.rs, dropping the four dead _output parameters that falsely
  implied output policy feeds derivation
- refresh batcher/token cost snapshots (small CU improvements from the
  finish_computed_handle dedup)
