--- solana/programs/zama-host/src/instructions/fhe_eval/handles.rs
@@ -1,172 +1,0 @@
-use super::*;
-
-pub(super) struct EvalHandleContext<'a> {
-    pub chain_id: u64,
-    pub previous_bank_hash: &'a [u8; 32],
-    pub unix_timestamp: i64,
-    pub context_id: &'a [u8; 32],
-}
-
-// Durable and instruction-local outputs derive the identical handle: the base
-// handle (op/operands/type + per-block entropy) already distinguishes distinct
-// ciphertexts, matching EVM `FHEVMExecutor`, which binds no per-output value.
-
-pub(super) fn expected_binary_eval_result(
-    op: FheBinaryOpCode,
-    lhs: [u8; 32],
-    rhs: [u8; 32],
-    scalar: bool,
-    output_fhe_type: u8,
-    handle_context: &EvalHandleContext<'_>,
-    op_index: u16,
-) -> [u8; 32] {
-    computed_eval_handle(
-        op,
-        lhs,
-        rhs,
-        scalar,
-        output_fhe_type,
-        handle_context.chain_id,
-        *handle_context.previous_bank_hash,
-        handle_context.unix_timestamp,
-        *handle_context.context_id,
-        op_index,
-    )
-}
-
-pub(super) fn expected_ternary_eval_result(
-    op: FheTernaryOpCode,
-    control: [u8; 32],
-    if_true: [u8; 32],
-    if_false: [u8; 32],
-    output_fhe_type: u8,
-    handle_context: &EvalHandleContext<'_>,
-    op_index: u16,
-) -> [u8; 32] {
-    computed_eval_ternary_handle(
-        op,
-        control,
-        if_true,
-        if_false,
-        output_fhe_type,
-        handle_context.chain_id,
-        *handle_context.previous_bank_hash,
-        handle_context.unix_timestamp,
-        *handle_context.context_id,
-        op_index,
-    )
-}
-
-pub(super) fn expected_trivial_eval_result(
-    plaintext: [u8; 32],
-    fhe_type: u8,
-    handle_context: &EvalHandleContext<'_>,
-    op_index: u16,
-) -> [u8; 32] {
-    computed_eval_trivial_handle(
-        plaintext,
-        fhe_type,
-        handle_context.chain_id,
-        *handle_context.previous_bank_hash,
-        handle_context.unix_timestamp,
-        *handle_context.context_id,
-        op_index,
-    )
-}
-
-pub(super) fn expected_rand_eval_seed(
-    handle_context: &EvalHandleContext<'_>,
-    op_index: u16,
-) -> [u8; 16] {
-    computed_eval_rand_seed(
-        handle_context.chain_id,
-        *handle_context.previous_bank_hash,
-        handle_context.unix_timestamp,
-        *handle_context.context_id,
-        op_index,
-    )
-}
-
-pub(super) fn expected_unary_eval_result(
-    op: FheUnaryOpCode,
-    operand: [u8; 32],
-    output_fhe_type: u8,
-    handle_context: &EvalHandleContext<'_>,
-    op_index: u16,
-    _output: &FheEvalOutput,
-) -> [u8; 32] {
-    computed_eval_unary_handle(
-        op,
-        operand,
-        output_fhe_type,
-        handle_context.chain_id,
-        *handle_context.previous_bank_hash,
-        handle_context.unix_timestamp,
-        *handle_context.context_id,
-        op_index,
-    )
-}
-
-pub(super) fn expected_sum_eval_result(
-    operand_handles: &[[u8; 32]],
-    fhe_type: u8,
-    handle_context: &EvalHandleContext<'_>,
-    op_index: u16,
-    _output: &FheEvalOutput,
-) -> [u8; 32] {
-    computed_eval_sum_handle(
-        operand_handles,
-        fhe_type,
-        handle_context.chain_id,
-        *handle_context.previous_bank_hash,
-        handle_context.unix_timestamp,
-        *handle_context.context_id,
-        op_index,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(super) fn expected_is_in_eval_result(
-    value_handle: [u8; 32],
-    set_handles: &[[u8; 32]],
-    fhe_type: u8,
-    handle_context: &EvalHandleContext<'_>,
-    op_index: u16,
-    _output: &FheEvalOutput,
-) -> [u8; 32] {
-    computed_eval_is_in_handle(
-        value_handle,
-        set_handles,
-        fhe_type,
-        handle_context.chain_id,
-        *handle_context.previous_bank_hash,
-        handle_context.unix_timestamp,
-        *handle_context.context_id,
-        op_index,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(super) fn expected_mul_div_eval_result(
-    factor1: [u8; 32],
-    factor2: [u8; 32],
-    scalar: bool,
-    divisor: [u8; 32],
-    output_fhe_type: u8,
-    handle_context: &EvalHandleContext<'_>,
-    op_index: u16,
-    _output: &FheEvalOutput,
-) -> [u8; 32] {
-    computed_eval_mul_div_handle(
-        factor1,
-        factor2,
-        divisor,
-        scalar,
-        output_fhe_type,
-        handle_context.chain_id,
-        *handle_context.previous_bank_hash,
-        handle_context.unix_timestamp,
-        *handle_context.context_id,
-        op_index,
-    )
-}
