+++ solana/runtime-tests/cost-snapshots/batcher_mollusk.json
@@ -1,6 +1,6 @@
 {
   "claim": {
-    "compute_units": 311116,
+    "compute_units": 311092,
     "unique_accounts": 22,
     "instruction_data_bytes": 8,
     "cpi_instructions": 12,
@@ -8,7 +8,7 @@
     "max_cpi_instruction_data_bytes": 1065
   },
   "dispatch": {
-    "compute_units": 303559,
+    "compute_units": 303535,
     "unique_accounts": 18,
     "instruction_data_bytes": 8,
     "cpi_instructions": 9,
@@ -16,7 +16,7 @@
     "max_cpi_instruction_data_bytes": 969
   },
   "join": {
-    "compute_units": 352949,
+    "compute_units": 352925,
     "unique_accounts": 20,
     "instruction_data_bytes": 223,
     "cpi_instructions": 14,
@@ -24,7 +24,7 @@
     "max_cpi_instruction_data_bytes": 1427
   },
   "open_batch": {
-    "compute_units": 218936,
+    "compute_units": 218924,
     "unique_accounts": 24,
     "instruction_data_bytes": 16,
     "cpi_instructions": 20,
@@ -32,7 +32,7 @@
     "max_cpi_instruction_data_bytes": 249
   },
   "redeem_claim": {
-    "compute_units": 353117,
+    "compute_units": 353093,
     "unique_accounts": 22,
     "instruction_data_bytes": 8,
     "cpi_instructions": 12,
@@ -40,7 +40,7 @@
     "max_cpi_instruction_data_bytes": 1065
   },
   "redeem_dispatch": {
-    "compute_units": 243560,
+    "compute_units": 243536,
     "unique_accounts": 18,
     "instruction_data_bytes": 8,
     "cpi_instructions": 9,
@@ -48,7 +48,7 @@
     "max_cpi_instruction_data_bytes": 969
   },
   "redeem_join": {
-    "compute_units": 345450,
+    "compute_units": 345426,
     "unique_accounts": 20,
     "instruction_data_bytes": 223,
     "cpi_instructions": 14,
@@ -56,7 +56,7 @@
     "max_cpi_instruction_data_bytes": 1427
   },
   "redeem_open_batch": {
-    "compute_units": 214437,
+    "compute_units": 214425,
     "unique_accounts": 24,
     "instruction_data_bytes": 16,
     "cpi_instructions": 20,
